Bump version to trigger rebuild and propagation of new arm64 binary
The current arm64 binary that is available in BioC 3.16 is
unfortunately dynamically linked to the libgfortran.5.dylib and
libquadmath.0.dylib libraries located in /opt/R/arm64/gfortran/lib/.
This is the location of Coudert's gfortran on kjohnson, the Bioconductor
arm64 build machine for BioC 3.16. Because of this, the current binary
won't load on an arm64 machine that doesn't have Coudert's gfortran.

Adjustments have been made on kjohnson to address this. From now on,
the binary packages produced by the machine should work on a user
machine that doesn't have Coudert's gfortran.